scaled done problem works with trig functions
The values in the equations are not meaningful, but they were scaled to all have a similar range, to test if the problem in the trig functions lied in the fact that variables were not normalized. This appears to be the cause of the problems encountered with the trig function on the earlier GEKKO tests.